Update rfcs repo link
The README process list was originally written in Qiskit/qiskit#666 and
was assuming it'd be part of the meta-repository. This readme included a
link to the repo to fork and create a PR. However, a separate repo was
used instead and this link was never updated. This commit corrects the
oversight and sets the link to the proper repo url.