Add feedback from terra issue
A new Terra issue Qiskit/qiskit-terra#4105 was opened that has partial
overlap with the work being done here. It had some good ideas that
would make the v2 interface a better fit for a world where there are
backends that do more than just run circuits (or pulse schedules). This
commit starts the process of integrating the ideas from that into the
larger refactor pending for the provider interface.