Update README to reflect current process
This commit updates the RFC readme to better explain the purpose and
scope of the repository and the RFC process. When this repository was
originally introduced it was done with the intent to have a forum to
dicsuss large design discussions for the qiskit project that spanned the
entire project that involved multiple development teams. However, this
never really caught hold and the development process and the teams have
changed significantly since then. This commit should hopefully make it
clear that RFCs can be used for any design discussion on any qiskit
project as long as the discussion requires a wider discussion at the
design stage. Features or major changes solely in a single component
should feel free to use this process to discuss the design up front.