Preliminary representation of rvalue classical expressions in Terra
This is a proposal for initial representation of classical expressions
in Terra.  It is simple bit-level operations only on effectively `Clbit`
and `ClassicalRegister`, and is deliberately designed to be separable
from the rest of Terra so it can be replaced in the future if necessary,
without painful decoupling of these objects from core Terra ones.

This proposal notably does not require any new public methods or
trackers on `QuantumCircuit`, `Clbit` or `ClassicalRegister`.  This
includes _not_ having a way to store a runtime expression in a variable
of `QuantumCircuit` for re-use, nor of using these expressions in the
parameters to gates.  These may come in the future, but are deliberately
excluded from prelim exploratory representations.