Preliminary representation of rvalue classical expressions in Terra (#30)
* Preliminary representation of rvalue classical expressions in Terra

This is a proposal for initial representation of classical expressions
in Terra.  It is simple bit-level operations only on effectively `Clbit`
and `ClassicalRegister`, and is deliberately designed to be separable
from the rest of Terra so it can be replaced in the future if necessary,
without painful decoupling of these objects from core Terra ones.

This proposal notably does not require any new public methods or
trackers on `QuantumCircuit`, `Clbit` or `ClassicalRegister`.  This
includes _not_ having a way to store a runtime expression in a variable
of `QuantumCircuit` for re-use, nor of using these expressions in the
parameters to gates.  These may come in the future, but are deliberately
excluded from prelim exploratory representations.

* Fix typos

Co-authored-by: Kevin Hartman <kevin@hart.mn>

* Fix typos

Co-authored-by: Kevin Krsulich <kevin@krsulich.net>

* Fix title

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

* Expand on alternatives

* Add long-term notes

* Clarify usage examples

* Add ExprVisitor discussion

This replaces the `resources` field; there is no need for this to be
manually tracked during construction, and the visitor is how general
double-dynamic dispatch should be handled.

* Remove confusing use of '!'

* Include details of Boolean overloads

* Prep for merge

Co-authored-by: Luciano Bello <bel@zurich.ibm.com>

---------

Co-authored-by: Kevin Hartman <kevin@hart.mn>
Co-authored-by: Kevin Krsulich <kevin@krsulich.net>
Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Luciano Bello <bel@zurich.ibm.com>