Variable-declaration and storage representation in Qiskit
This is a proposal and design comments on how to begin adding the
representations of declared classical variables and stores to Qiskit,
including input variables, which correspond to the `input` keyword in
OpenQASM 3, and can be used to define parametric input of arbitrary
types into Qiskit.

For now, this RFC does not propose expanding Qiskit's classical type
system, so only `Bool` and `Uint` remain valid, and these parameters
cannot yet be used in gate-argument positions.  This is still strongly
desired, just not for this RFC.